Make sure example handlers use a structured event type
Internally the request event is effectively deserialized using:

  let event = serde_json::from_slice<T>(&body)

where `T` is the type used for the handler's event type. If this is a
String then Serde will fail to deserialize the JSON input (it won't ever
just be a string).

Handler's should use a structured type for the event that Serde will
be able to deserialize the incoming JSON event data into, so this
updates the examples to use serde_json::Value.