Corrected --target name in README, added links.
The text was referring to `x86_64-linux` as a target, which is not even a valid value. From what I know, the only valid target is `x86_64-unknown-linux-musl`, so we may just as well say so. I lost at least an hour (novice!) trying to figure it out. 

Relates to issue #216.