lambda-extension: remove logging
When running on AWS lambda I encountered the following error:

```
thread 'main' panicked at 'Could not determine the UTC offset on this
system. Possible causes are that the time crate does not implement
"local_offset_at" on your system, or that you are running in a
multi-threaded environment and the time crate is returning "None" from
"local_offset_at" to avoid unsafe behaviour. See the time crate's
documentation for more information.
(https://time-rs.github.io/internal-api/time/index.html#feature-flags):
IndeterminateOffset',
/home/simon/.cargo/registry/src/github.com-1ecc6299db9ec823/simple_logger-1.16.0/src/lib.rs:409:85
```

This means that the simple logger setup could not find the current time.
Instead, we can print to stdout which is also collected by CloudWatch.
We can leave more sophisticated control of the output to the user.