Expose the raw http path coming from the lambda event.
The API GW url includes th API GW stage information, so
it's not easy to know what's the raw http path coming into
the event.

With this change, we expose the raw http path in a general
way, so people that need to know the exact raw path have
a common interface, regardless of where the event comes from.

Signed-off-by: David Calavera <david.calavera@gmail.com>