Don't require Lambda `Service` to use `lambda_http::Error`
Currently, the `Service` that you pass to `lambda_http::run` is required
to use `lambda_http::Error` as its associated error type, a type-erased
boxed error type. However, all the runtime needs is for the error type
to implement `Debug` and `Display`. This commit relaxes this constraint,
allowing users to run a broader class of `Service`s.