Pin aws_lambda_events dependency to patch versions
If you already have a version of aws_lambda_events in your project, and you update lambda_http, the events package's version might not be updated. This can cause compilation issues like the ones we saw in #471. With this change, the events package will be update with new patch versions, which should prevent issues like that from happening.