Update Rust edition to 2021
The AWS SDK already made the upgrade to 2021. We've had the `warn_2018_idioms` lint flag enabled for a while.

Since we're making a breaking change in the next version of lambda_http, I thought it could be a good opportunity to get this in, and bump the version to all the crates at the same time.

Signed-off-by: David Calavera <david.calavera@gmail.com>