fix lambda-extension README (#511)
the current example code doesn't work if you copy/paste it into a repo created with
`cargo lambda new`, even if you change the `Cargo.toml` to pull in the `lambda-extension`
crate.  The default template doesn't bring in `tracing` with the `ansi` feature.  This
updates the README to have code that's more in line with the default template.