Add `with_request_context` to `RequestExt` (#523)
The trait `RequestExt` already includes functions to add query string
parameters, path parameters and stage variables to a request. This
commit adds another function to allow adding a `RequestContext` to the
request. This is useful in cases where a lambda function is called by
the API Gateway with a custom authorizer that adds parameters to the
`RequestContext`.

Related to #522

Co-authored-by: Simon Gasse <sgasse@users.noreply.github.com>