Bump patch version of lambda-http (#556)
This commit updates the version of `lambda-http` to 0.7.1. The crate now
depends on `aws-lambda-events` 0.7.2 or higher whereas it previously
depended on that of ^0.7. This crate currently does not compile against
`aws-lambda-events` 0.7.0 or 0.7.1. In `aws-lambda-events` 0.7.0, the module
`query_map` is private when it should be public for `lambda-http` to compile.
Similarly, in `aws-lambda-events` 0.7.1, the trait bound `QueryMap: FromStr`
is not satisfied so `lambda-http` does not compile either.

Co-authored-by: Saito <awsaito@c889f3b5ddc4.ant.amazon.com>