Implement custom deserializer for LambdaRequest
This deserializer gives us full control over the error message that we return for invalid payloads. The default message that Serde returns is usually very confusing, and it's been reported many times as something people don't understand.

This code is a copy of the code that Serde generates when it expands the Deserialize macro.

Signed-off-by: David Calavera <david.calavera@gmail.com>