Fix APIGW path with stage (#669)
* Fix APIGW path with stage

APIGW HTTP has started adding the stage to the path in the event. This change checks if the stage is already a prefix in the path, and skips adding it if so.

Signed-off-by: David Calavera <david.calavera@gmail.com>

* Add env variable to shortcircuit stage behavior.

There might be cases when you don't want the runtime to do anything with paths and stages. By setting AWS_LAMBDA_HTTP_IGNORE_STAGE_IN_PATH in the environment, we ignore this behavior completely.

Signed-off-by: David Calavera <david.calavera@gmail.com>

---------

Signed-off-by: David Calavera <david.calavera@gmail.com>