Allow running handler with a callback
## What?
This is a pretty naive approach that allows the end user to supply a callback that will be invoked once a single event loop has been finished.

## Why?
Because currently there is no good way(that I know of) to flush logs/traces/metrics. By having a callback that is called after all event handling is finished allows the end user to do a cleanup at the end of each lambda invocation.