Refactor request deserialization in lambda-http
Identifying different Lambda events that pertain to HTTP requests is notoriously
hard since they don't contain a simple `kind` property to help out. This is generally
true - not just a problem in Rust.

Up to 0.8 of lambda-http, the determination was made by using a private deserializer
available in the serde crate to deserialize.

In 0.9 this was modified to but far more clean, deserializing into a serde_json::RawValue
to check the general structure of the request JSON, then deserializing the contained
string slice into the types.

Both approaches attempted a full deserialization of each enabled request type in
turn, trying the next on failure of the previous.

The additional downsides of the 0.9 approach are:

* The attempted deserializations use the original string rather than the partially parsed-data
* The unintended consequence of disabling the possibility of using simd_json as the
  incoming deserializer - not exactly a supported feature anyway, but simd_json has a problem
  with RawValue.

This commit attempts to solve all these issues by:

* Deserializing into a serde_json::Value using the incoming deserializer
* Testing for "type determinants" in the Value instead of trying full deserialization
* Completing the deserialization to the discovered type from the Value.