Support simd_json as an optional alternative to serde_json
The approach is to abstract _all_  original serde_json references to a new lib crate
in the workspace - json-impl.

That crate as a feature swicth between serde_json and simd_json. In

The initial phase, is to to get all the other crates to depend on json-impl with the serde feature,
instead of serde_json directly, so in the end nothing should change.

The next phase is to activate the simd feature and see what fails, then start mitigating
the failures.