Start hacking stuff into mutable slices
The _problem_ with simd_json is that it requires _mutable_  slices to operate while serde_json does not.

This FIRST PASS (certainly not the end solution) is to provide &mut adapter of the from_slice function for serde_json,
a pseudo-safe adapter of from_str for simd_json, and then change all the places where stuff is flagged as needing to
be mutable to be so - MOSTLY, however, this is in TESTS -  so eventually we'll clean that up.