We're down to the core of the issue
We've removed Deserialize from of lambda_http::request::LambdaRequest ...
that makes EVERYTHING compile in the simd_json context ... with the exception
of th calls to lambda_runtime::run which require the expected request type to be
serde::Deserialize.

So now it's back to that issue - which, if I recall well, was the blocker back when I
first started this!